refactor(globe): decouple funding request parsers
